# backend/swagger.yaml
openapi: 3.0.0
info:
  title: User & Run Management API
  version: 1.0.0
  description: API für Benutzerverwaltung und Laufdaten
paths:
  /login:
    post:
      summary: User login
      description: Authenticate a user and return access and refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: "securepassword123"
      responses:
        200:
          description: Successfully authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI..."
                  refreshToken:
                    type: string
                    example: "d1b2c3d4e5f6g7h8i9j0..."
        401:
          description: Invalid credentials
  /token:
    post:
      summary: Refresh access token
      description: Generate a new access token using a valid refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  example: "d1b2c3d4e5f6g7h8i9j0..."
      responses:
        200:
          description: New access token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI..."
        401:
          description: Invalid or missing refresh token
        403:
          description: Refresh token is no longer valid
  /logout:
    delete:
      summary: User logout
      description: Invalidate a refresh token to log out the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  example: "d1b2c3d4e5f6g7h8i9j0..."
      responses:
        204:
          description: Successfully logged out
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT